Add update_reorder_diff/2
Reorder was in category of "changed", but reorder-ing fmodels get
too much data on wire for lots of fmodels. We separate reorder as
"reorder" to not sending fmodel.sch.

We think it's safer(consistency) to send ALL items' indices rather
than only sending reorered items because it's store order in each
record without uniqueness guarantee, it may get out of sync.